refactor(electron): name the shared main-process logger app-logger
`src/host-proxy/executors/logger.ts` already default-exports a `log`, and it
means the opposite thing: an injectable seam that forwards to whichever logger
the client installed, rather than a logger of its own. Two modules named
`logger` in one package, both default-exporting `log`, read as one concept.

Also drops the test's listener-capture teardown, since the suite runs one
process per file, and the type assertions that teardown needed.
