+++ clients/macos/src/main/logger.ts
@@ -1,1 +1,4 @@
-export { default, getLogFilePaths } from "@vellumai/electron-desktop/logger";
+export {
+  default,
+  getLogFilePaths,
+} from "@vellumai/electron-desktop/app-logger";
